Crate rust_gh_client

Overview

This program will show you your publicly available GitHub info. It might also lightly roast your repositories.

Installation

Simply download a precompiled binary from our GitHub repo and run the program in a terminal. If this isn’t to your liking or we didn’t build for your platform, you can always clone/download the source code from the repo and try building it yourself. This will require installing cargo, which is actually quite simple. Follow this guide and, once cargo is installed, run

cargo run

in your terminal from anywhere within the cloned repository.

Usage

This program takes no arguments. Simply run with with cargo run or download one of our binaries and run it directly in the terminal. Enter your GitHub username, and you will be prompted from there.

Modules

roasts

types

Constants

SEPARATOR

Functions

main
This is the entrypoint for the whole program. Takes no arguments. Prompts the user for their GitHub login and checks to see if it exists on the web. If it does not, the user will be re-prompted, and if it does, continues to the main menu.

main_menu
Presents the user with the main menu, which allows them to see their own info, info on their repos, sign in as a different user, or quit.

print_all_user_info
Prints all of the user’s public Github information as long as a value for the field exists.

print_repo_info
Prints a repo’s information along with roasts associated with certain repo fields.

print_repo_titles
Prints the names of each repo for all user’s repos in order of creation.

print_repos_menu
Presents the user with the menu with options about their repos. This is also the function that fetches the repos from the web, so it can return an error if that goes wrong for whatever reason. Fetches up to 40 repos.

print_user_info
Presents the user with the user’s information menu. It will show some general information about the user and then offer the options to print all public information about the user or to return to the main menu.

user_exists
Checks the GH API to see if a user with the login in uname exists. Returns true iff the response is successful (200), so this could return false for issues other than the user not existing, such as a failed connection or a non-404 error.