Module rust_gh_client::roasts

Constants

DAY_IN_SECONDS

HOUR_IN_SECONDS

MINUTE_IN_SECONDS

ROUGH_MONTH_IN_SECONDS

YEAR_IN_SECONDS

Functions

roast_default_branch

Prints a message iff the given string is equal to master. Should be called with a repo object’s default_branch field from the GitHub API.

roast_fork

Prints a message if the given bool is true. It should represent whether or not a repo is a fork.

roast_issues

Prints a message depending on how many issues a repo has.

roast_language

Prints a message depending on the given string, which should be the language field of an API repo object. Will still print a message if the field is None or if the language is not one of the ones I wrote a specific message for.

roast_license

Prints a message depending on the license object of a repo. Will still print a generic message if the license or it’s name is None.

roast_num_repos

Prints some messages depending on the number of public repositories found when querying their repos endpoint

roast_stars

Prints a message depending on the number of stars a repo has

roast_updated_at

Prints a message depending on how long ago the given date is. Will still print a generic message if the date can’t be parsed.